requirements.txt

rdflib==6.1.1
asgiref==3.2.10
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
bagit==1.8.1
beautifulsoup4==4.8.2
biopython==1.79
blinker==1.4
CacheControl==0.11.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
-e git+https://github.com/etal/cnvkit@09208296f95685bc7a34660d440a83aaa14964b8#egg=CNVkit
colorama==0.4.3
coloredlogs==7.3
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
cwltool==2.0.20200224214940
cycler==0.11.0
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.7.0rc1
distro==1.7.0
Django==3.0.8
django-cors-headers==3.11.0
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.17.3
fonttools==4.29.1
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.14.0
humanfriendly==4.18
hyperlink==19.0.0
idna==2.10
importlib-resources==5.4.0
incremental==16.10.1
isodate==0.6.0
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==0.17.0
jsonschema==4.4.0
jupyter-core==4.9.2
keyring==18.0.1
kiwisolver==1.3.2
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.5.0
MarkupSafe==2.1.0
matplotlib==3.5.1
mistune==0.8.4
msgpack==0.6.2
mypy-extensions==0.4.3
nbformat==5.2.0
netifaces==0.10.4
networkx==2.4
numpy==1.22.2
oauthlib==3.1.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pathlib2==2.3.7.post1
pbr==5.8.1
pexpect==4.6.0
Pillow==9.0.1
pomegranate==0.14.8
prov==1.5.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycparser==2.20
pydot==1.4.2
pydotplus==2.0.2
pyfaidx==0.6.4
Pygments==2.11.2
PyGObject==3.42.0
PyHamcrest==2.0.3
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
pysam==0.18.0
python-dateutil==2.8.2
python-debian==0.1.43
python3-openid==3.2.0
pytz==2020.1
pyufw==0.0.3
PyVCF==0.6.8
PyYAML==6.0
pyzmq==22.3.0

rdflib-jsonld==0.5.0
reportlab==3.6.8
requests==2.25.0
requests-oauthlib==1.3.0
requests-unixsocket==0.3.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.6
schema-salad==5.0.20200416112825
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==3.3.1
service-identity==21.1.0
shellescape==3.4.1
simplejson==3.17.6
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sos==0.22.7
soupsieve==2.3.1
SPARQLWrapper==2.0.0
sqlparse==0.3.1
ssh-import-id==5.11
systemd-python==234
testresources==2.0.1
threadpoolctl==3.1.0
tqdm==4.63.0
traitlets==5.1.1
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.2
wadllib==1.3.6
webencodings==0.5.1
whitenoise==3.3.1
XlsxWriter==3.0.3
xxhash==3.0.0
zipp==3.7.0
psycopg2==2.8.5
psycopg2-binary==2.8.5